icn_reputation::trust_graph

Struct TrustGraph

pub struct TrustGraph { /* private fields */ }

Core trust graph structure representing the network of trust relationships

Implementations

impl TrustGraph

pub fn new() -> Self

Create a new empty trust graph

pub fn add_edge(&mut self, edge: TrustEdge)

Add or update a trust edge in the graph

pub fn remove_edge(&mut self, from: &Did, to: &Did) -> Option<TrustEdge>

Remove a trust edge from the graph

pub fn get_edge(&self, from: &Did, to: &Did) -> Option<&TrustEdge>

Get a trust edge between two entities

pub fn get_outgoing_edges(&self, from: &Did) -> Option<&HashMap<Did, TrustEdge>>

Get all outgoing edges from an entity

pub fn get_incoming_edges(&self, to: &Did) -> Option<&HashMap<Did, TrustEdge>>

Get all incoming edges to an entity

pub fn get_all_nodes(&self) -> Vec<Did>

Get all entities (DIDs) in the graph

pub fn edge_count(&self) -> usize

Get the number of edges in the graph

pub fn node_count(&self) -> usize

Get the number of nodes in the graph

pub fn cleanup_expired_edges( &mut self, time_provider: &dyn TimeProvider, max_age_seconds: u64, )

Clean up expired edges based on a time provider

pub fn invalidate_cache(&mut self)

Clear all cached scores

pub fn get_cached_score( &self, did: &Did, max_cache_age: u64, current_time: u64, ) -> Option<f64>

Get a cached trust score if available and not stale

pub fn cache_score(&mut self, did: Did, score: f64, timestamp: u64)

Cache a computed trust score